fixed merge conflicts
