+++ rareapi/views/__init__.py
@@ -1,4 +1,5 @@
 from .auth import register_user, login_user
 from .tag_view import TagView
 from .posts_view import PostView
-from .categories_view import CategoryView
+from .category_view import CategoryView
+
